locations ala equipment
